requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.0.0
Django==4.2
django-cors-headers==4.5.0
djangorestframework==3.14.0
drf-yasg==1.21.8
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.0
packaging==24.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
service-identity==24.2.0
setuptools==75.2.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
zope.interface==7.1.1
